Signed-off-by: R. Timothy Edwards <tim@opencircuitdesign.com>
Added scripts to magic for generated devices, as well
as an overhaul of the tech file to split it into
multiple files.  The generated devices have also been
split into multiple files with one file per device
class.  Certain devices that have fixed layout are
kept as a Tcl script of magic commands which build
the cell.  The magicrc startup file now sources the
generated devices.